+++ core/src/findlib/bfile.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2003-2010 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -376,12 +376,9 @@ extern "C" HANDLE get_osfhandle(int fd);
 
 void binit(BareosWinFilePacket* bfd)
 {
-  memset(bfd, 0, sizeof(BareosWinFilePacket));
+  new (bfd) BareosWinFilePacket();
   bfd->fid = -1;
-  bfd->mode = BF_CLOSED;
   bfd->use_backup_api = have_win32_api();
-  bfd->cmd_plugin = false;
-  bfd->pvContext = NULL;
 }
 
 /**