+++ core/src/stored/btape.cc
@@ -57,6 +57,22 @@
 #include "lib/watchdog.h"
 #include "include/jcr.h"
 
+inline void read_with_check(int fd, void* buf, size_t count)
+{
+  if (read(fd, buf, count) < 0) {
+    BErrNo be;
+    Emsg1(M_FATAL, 0, _("read failed: %s\n"), be.bstrerror());
+  }
+}
+
+inline void write_with_check(int fd, void* buf, size_t count)
+{
+  if (write(fd, buf, count) < 0) {
+    BErrNo be;
+    Emsg1(M_FATAL, 0, _("write failed: %s\n"), be.bstrerror());
+  }
+}
+
 namespace storagedaemon {
 extern bool ParseSdConfig(const char* configfile, int exit_code);
 }
@@ -278,10 +294,10 @@ int main(int margc, char* margv[])
       if (bstrcmp(director->resource_name_, DirectorName)) { break; }
     }
     if (!director) {
-      Emsg2(
-          M_ERROR_TERM, 0,
-          _("No Director resource named %s defined in %s. Cannot continue.\n"),
-          DirectorName, configfile);
+      Emsg2(M_ERROR_TERM, 0,
+            _("No Director resource named %s defined in %s. Cannot "
+              "continue.\n"),
+            DirectorName, configfile);
     }
   }
 
@@ -428,7 +444,7 @@ static void FillBuffer(fill_mode_t mode, char* buf, uint32_t len)
     case FILL_RANDOM:
       fd = open("/dev/urandom", O_RDONLY);
       if (fd != -1) {
-        read(fd, buf, len);
+        read_with_check(fd, buf, len);
         close(fd);
       } else {
         uint32_t* p = (uint32_t*)buf;
@@ -2274,15 +2290,15 @@ static void fillcmd()
   sprintf(buf, "%s/btape.state", working_directory);
   fd = open(buf, O_CREAT | O_TRUNC | O_WRONLY, 0640);
   if (fd >= 0) {
-    write(fd, &btape_state_level, sizeof(btape_state_level));
-    write(fd, &simple, sizeof(simple));
-    write(fd, &last_block_num1, sizeof(last_block_num1));
-    write(fd, &last_block_num2, sizeof(last_block_num2));
-    write(fd, &last_file1, sizeof(last_file1));
-    write(fd, &last_file2, sizeof(last_file2));
-    write(fd, last_block1->buf, last_block1->buf_len);
-    write(fd, last_block2->buf, last_block2->buf_len);
-    write(fd, first_block->buf, first_block->buf_len);
+    write_with_check(fd, &btape_state_level, sizeof(btape_state_level));
+    write_with_check(fd, &simple, sizeof(simple));
+    write_with_check(fd, &last_block_num1, sizeof(last_block_num1));
+    write_with_check(fd, &last_block_num2, sizeof(last_block_num2));
+    write_with_check(fd, &last_file1, sizeof(last_file1));
+    write_with_check(fd, &last_file2, sizeof(last_file2));
+    write_with_check(fd, last_block1->buf, last_block1->buf_len);
+    write_with_check(fd, last_block2->buf, last_block2->buf_len);
+    write_with_check(fd, first_block->buf, first_block->buf_len);
     close(fd);
     Pmsg2(0, _("Wrote state file last_block_num1=%d last_block_num2=%d\n"),
           last_block_num1, last_block_num2);
@@ -2341,15 +2357,15 @@ static void unfillcmd()
   fd = open(buf, O_RDONLY);
   if (fd >= 0) {
     uint32_t state_level;
-    read(fd, &state_level, sizeof(btape_state_level));
-    read(fd, &simple, sizeof(simple));
-    read(fd, &last_block_num1, sizeof(last_block_num1));
-    read(fd, &last_block_num2, sizeof(last_block_num2));
-    read(fd, &last_file1, sizeof(last_file1));
-    read(fd, &last_file2, sizeof(last_file2));
-    read(fd, last_block1->buf, last_block1->buf_len);
-    read(fd, last_block2->buf, last_block2->buf_len);
-    read(fd, first_block->buf, first_block->buf_len);
+    read_with_check(fd, &state_level, sizeof(btape_state_level));
+    read_with_check(fd, &simple, sizeof(simple));
+    read_with_check(fd, &last_block_num1, sizeof(last_block_num1));
+    read_with_check(fd, &last_block_num2, sizeof(last_block_num2));
+    read_with_check(fd, &last_file1, sizeof(last_file1));
+    read_with_check(fd, &last_file2, sizeof(last_file2));
+    read_with_check(fd, last_block1->buf, last_block1->buf_len);
+    read_with_check(fd, last_block2->buf, last_block2->buf_len);
+    read_with_check(fd, first_block->buf, first_block->buf_len);
     close(fd);
     if (state_level != btape_state_level) {
       Pmsg0(-1, _("\nThe state file level has changed. You must redo\n"
@@ -2529,9 +2545,8 @@ static bool do_unfill()
     goto bail_out;
   }
   if (CompareBlocks(last_block, block)) {
-    Pmsg0(
-        -1,
-        _("\nThe last block on the second tape matches. Test succeeded.\n\n"));
+    Pmsg0(-1, _("\nThe last block on the second tape matches. Test "
+                "succeeded.\n\n"));
     rc = true;
   }
 
@@ -2811,7 +2826,8 @@ static struct cmdstruct commands[] = {
     {NT_("scanblocks"), scan_blocks,
      _("Bareos read block by block to EOT and report")},
     {NT_("speed"), speed_test,
-     _("[file_size=n(GB)|nb_file=3|skip_zero|skip_random|skip_raw|skip_block] "
+     _("[file_size=n(GB)|nb_file=3|skip_zero|skip_random|skip_raw|skip_block]"
+       " "
        "report drive speed")},
     {NT_("status"), statcmd, _("print tape status")},
     {NT_("test"), testcmd, _("General test Bareos tape functions")},
@@ -2861,7 +2877,8 @@ static void usage()
       stderr,
       _("Usage: btape <options> <device_name>\n"
         "       -b <file>     specify bootstrap file\n"
-        "       -c <path>     specify Storage configuration file or directory\n"
+        "       -c <path>     specify Storage configuration file or "
+        "directory\n"
         "       -D <director> specify a director name specified in the "
         "Storage\n"
         "                     configuration file for the Key Encryption Key "