+++ core/src/stored/sd_cmds.cc
@@ -2,7 +2,7 @@
    BAREOS - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2012 Planets Communications B.V.
-   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -60,13 +60,13 @@ static bool StartReplicationSession(JobControlRecord* jcr);
 static bool ReplicateData(JobControlRecord* jcr);
 static bool EndReplicationSession(JobControlRecord* jcr);
 
-struct s_cmds {
+struct s_sd_cmds {
   const char* cmd;
   bool (*func)(JobControlRecord* jcr);
 };
 
 // The following are the recognized commands from the Remote Storage daemon
-static struct s_cmds sd_cmds[] = {
+static struct s_sd_cmds sd_cmds[] = {
     {"start replicate", StartReplicationSession},
     {"replicate data", ReplicateData},
     {"end replicate", EndReplicationSession},